different class styles

+++ test/any/class1.cpp
@@ -1,9 +1,9 @@
 
 #include "kul/gpu.hpp"
 
-static constexpr size_t WIDTH = 1024, HEIGHT = 1024;
-static constexpr size_t NUM = WIDTH * HEIGHT;
-static constexpr size_t THREADS_PER_BLOCK_X = 16, THREADS_PER_BLOCK_Y = 16;
+static constexpr uint32_t WIDTH = 1024, HEIGHT = 1024;
+static constexpr uint32_t NUM = WIDTH * HEIGHT;
+static constexpr uint32_t THREADS_PER_BLOCK_X = 16, THREADS_PER_BLOCK_Y = 16;
 
 template<typename Float, bool GPU = false>
 struct DevClass : kul::gpu::DeviceClass<GPU>
@@ -50,15 +50,15 @@ __global__ void vectoradd(GPUClass<T>* a, GPUClass<T> const* b, GPUClass<T> cons
 }
 
 template<typename Float>
-size_t test(){
+uint32_t test(){
   std::vector<Float> hostB(NUM), hostC(NUM);
-  for (size_t i = 0; i < NUM; i++) hostB[i] = i;
-  for (size_t i = 0; i < NUM; i++) hostC[i] = i * 100.0f;
+  for (uint32_t i = 0; i < NUM; i++) hostB[i] = i;
+  for (uint32_t i = 0; i < NUM; i++) hostC[i] = i * 100.0f;
   DevClass<Float> devA(NUM), devB(hostB), devC(hostC);
   kul::gpu::Launcher{WIDTH, HEIGHT, THREADS_PER_BLOCK_X, THREADS_PER_BLOCK_Y}(
       vectoradd<Float>, devA(), devB(), devC());
   auto hostA = devA.data();
-  for (size_t i = 0; i < NUM; i++)
+  for (uint32_t i = 0; i < NUM; i++)
     if (hostA[i] != (hostB[i] + hostC[i])) return 1;
   return 0;
 }
